+++ selene/support/conditions/have.py
@@ -119,32 +119,32 @@ def tag_containing(name: str) -> Condition[Element]:
 # *** SeleneCollection conditions ***
 
 
-def size(number: int) -> Condition[Collection]:
-    return match.collection_has_size(number)
+def size(number: int | dict):
+    return match.size(number)
 
 
-def size_less_than(number: int) -> Condition[Collection]:
-    return match.collection_has_size_less_than(number)
+def size_less_than(number: int):
+    return match.size_less_than(number)
 
 
-def size_less_than_or_equal(number: int) -> Condition[Collection]:
-    return match.collection_has_size_less_than_or_equal(number)
+def size_less_than_or_equal(number: int):
+    return match.size_less_than_or_equal(number)
 
 
-def size_greater_than(number: int) -> Condition[Collection]:
-    return match.collection_has_size_greater_than(number)
+def size_greater_than(number: int):
+    return match.size_greater_than(number)
 
 
-def size_at_least(number: int) -> Condition[Collection]:
+def size_at_least(number: int):
     warnings.warn(
         'might be deprecated; use have.size_greater_than_or_equal instead',
         PendingDeprecationWarning,
     )
-    return match.collection_has_size_greater_than_or_equal(number)
+    return match.size_greater_than_or_equal(number)
 
 
-def size_greater_than_or_equal(number: int) -> Condition[Collection]:
-    return match.collection_has_size_greater_than_or_equal(number)
+def size_greater_than_or_equal(number: int):
+    return match.size_greater_than_or_equal(number)
 
 
 # TODO: consider accepting ints