+++ tests/integration/shared_browser/browser__config__wait_decorator_with_decorator_from_support_logging_test.py
@@ -102,7 +102,7 @@ Message:\u0020
 Timed out after 0.3s, while waiting for:
 browser.all(('css selector', '#todo-list>li')).have texts ['a', 'b', 'c']
 
-Reason: ConditionMismatch: actual visible texts: ['a', 'c']\n
+Reason: ConditionMismatch: actual: ['a', 'c']\n
             '''.strip()
             in handler.stream
         )